Update FootprintOrderFlow.cs
Option to align the aggregation levels to the session open, creating even levels across all bars.,
Fixed Predator signals that appeared but didn't disappear as the bar developed, and the  signal is no longer valid.,
Improved Styling of Value Area and POC.,
Option to show/hide the Signal Legend.,
Added additional plots to pinpoint the position of the POC level compared to the Value Area and the Bar. For use in strategies.,
ToDo: Additional signals. 
  Stopping Volume
  Delta Flip